+++ CHANGELOG.md
@@ -76,7 +76,99 @@ broker.createService({
 broker.call("staging.test.hello");
 ```
 
+## Module log level configuration is removed
+The module log level is not supported. The `logLevel` option can be only `String`. It is used if the logger is the `console`. **In case of external loggers you have to handle log levels.**
+
 # New 
+
+## Better logging [#61](https://github.com/ice-services/moleculer/issues/61)
+The whole Moleculer logger is rewritten. It supports better the external loggers. Changed the built-in log message format.
+
+### Built-in `console` logger
+```js
+const broker = createBroker({ 
+    logger: console, 
+    logLevel: "debug"
+});
+```
+New console output:
+![image](https://user-images.githubusercontent.com/306521/29127309-011bd0e0-7d21-11e7-87e2-c2d83352a857.png)
+
+**Console logger with custom `logFormatter`**
+```js
+const broker = new ServiceBroker({ 
+    logger: console, 
+    logFormatter(level, args, bindings) {
+        return level.toUpperCase() + " " + bindings.nodeID + ": " + args.join(" ");
+    }
+});
+broker.logger.warn("Warn message");
+broker.logger.error("Error message");
+```
+Output:
+```
+WARN norbi-pc2: Warn message
+ERROR norbi-pc2: Error message
+```
+
+### External loggers
+
+**[Pino](http://getpino.io/)**
+```js
+const pino = require("pino")({ level: "debug" });
+const broker = new ServiceBroker({ 
+    logger: bindings => pino.child(bindings)
+});
+```
+
+Output:
+![image](https://user-images.githubusercontent.com/306521/29127258-e151e3bc-7d20-11e7-9995-025f53cf41ec.png)
+
+
+**[Bunyan](https://github.com/trentm/node-bunyan)**
+```js
+const bunyan = require("bunyan");
+const logger = bunyan.createLogger({ name: "moleculer", level: "debug" });
+const broker = new ServiceBroker({ 
+    logger: bindings => logger.child(bindings)
+});
+```
+
+Output:
+![image](https://user-images.githubusercontent.com/306521/29127286-f2203428-7d20-11e7-84f1-c81aeaaaef53.png)
+
+**[Winston](https://github.com/winstonjs/winston)**
+```js
+const broker = new ServiceBroker({ 
+    logger: bindings => new winston.Logger({
+        transports: [
+            new (winston.transports.Console)({
+                timestamp: true,
+                colorize: true,
+                prettyPrint: true
+            })
+        ]
+    })
+});
+```
+
+**[Winston context](https://github.com/citrix-research/node-winston-context)**
+```js
+const WinstonContext = require("winston-context");
+const winston = require("winston");
+const broker = createBroker({ 
+    logger: bindings => extend(new WinstonContext(winston, "", bindings))
+});
+```
+
+The `bindings` contains the following properties:
+- `ns` - namespace
+- `nodeID` - nodeID
+- `mod` - type of core module: `broker`, `cacher`, `transit`, `transporter`
+- `svc` - service name
+- `ver` - service version
+
+
 
 ## Dynamic service load & destroy
 Available to load & destroy services after the broker started. For example you can hot-reload your services in runtime. The remote nodes will be notified about changes. The broker will emit a `services.changed` event locally.